+++ Userland/Libraries/LibWeb/DOM/AbstractRange.idl
@@ -1,0 +1,8 @@
+[Exposed=Window]
+interface AbstractRange {
+    readonly attribute Node startContainer;
+    readonly attribute unsigned long startOffset;
+    readonly attribute Node endContainer;
+    readonly attribute unsigned long endOffset;
+    readonly attribute boolean collapsed;
+};
