+++ Userland/Libraries/LibWeb/DOM/Range.idl
@@ -1,17 +1,12 @@
 #import <DOM/Node.idl>
+#import <DOM/AbstractRange.idl>
 
-interface Range {
+[Exposed=Window]
+interface Range : AbstractRange {
 
     constructor();
 
-    readonly attribute boolean collapsed;
-
     readonly attribute Node commonAncestorContainer;
-
-    readonly attribute Node startContainer;
-    readonly attribute unsigned long startOffset;
-    readonly attribute Node endContainer;
-    readonly attribute unsigned long endOffset;
 
     undefined setStart(Node node, unsigned long offset);
     undefined setEnd(Node node, unsigned long offset);
